refactor([issue-741]): extract autonomous-job scheduler from cos.js
Final slice of the cos.js god-file split. Move the autonomous-job and
improvement-check timer machinery — computeNextJobFireTime,
registerSingleJobSchedule, registerJobSchedules, unregisterJobSchedules,
executeScheduledJob with its spawning-guard (spawningJobIds / addSpawningJob /
clearSpawningJob), and scheduleNextImprovementCheck — out of cos.js into a
self-contained cosJobScheduler.js. cos.js drops to ~930 lines (from 3115
originally), now just daemon lifecycle + the task-spawn scheduler + init wiring
+ re-exports.

The new module imports only sibling services (no import back to cos.js): the
paused check reads loadState() directly (was cos.js's isPaused), and the
improvement-check timer requests a dequeue via the new `cos:dequeue-requested`
event instead of calling cos.js's dequeueNextTask — init() wires that event to
dequeueNextTask, mirroring the pre-extraction direct call. cos.js imports the
five entry points it drives (registerJobSchedules, unregisterJobSchedules,
scheduleNextImprovementCheck, registerSingleJobSchedule, clearSpawningJob) for
start()/stop()/init() and the job-lifecycle handlers.

Verified with a real `node --input-type=module -e import()` link check on both
modules; full suite green (9602 tests). Faithful relocation — no behavior change.

Closes #741
